Remove matrices support + split sequences wrappers
