+++ src/count_kmers_specific/count_kmers_specific.h
@@ -1,6 +1,3 @@
 #pragma once
 
-#include "count_kmers_integer_matrix.h"
 #include "count_kmers_string_vector.h"
-#include "count_kmers_numeric_matrix.h"
-#include "count_kmers_string_matrix.h"
