--- src/count_kmers_specific/rmatrix_row_getter.h
@@ -1,20 +1,0 @@
-#pragma once
-
-#include <Rcpp.h>
-// [[Rcpp::depends(RcppParallel)]]
-//' @importFrom  RcppParallel RcppParallelLibs
-#include<RcppParallel.h>
-#include <functional>
-#include <memory>
-#include <vector>
-#include "../kmer_task_config.h"
-
-template<class rcpp_matrix_t, class elem_t>
-inline SequenceGetter_t<typename RcppParallel::RMatrix<elem_t>::Row>
-getRMatrixRowGetter(rcpp_matrix_t &rcppMatrix, int rowOffset = 0) {
-    RcppParallel::RMatrix<elem_t> wrappedMatrix(rcppMatrix);
-    return [wrappedMatrix, rowOffset](int rowNum) -> typename RcppParallel::RMatrix<elem_t>::Row {
-        return wrappedMatrix.row(rowNum + rowOffset);
-    };
-}
-
